Add files via upload
Read the data, group them by the gender indicator, you will now have two sets of data, one with males, one with females. 
A single set should not contain duplicates, From the example data above, Jason should only be in the male set once, however Billy will be in the male set as well as the female set.
Run the good match program for every entry in the first set against every entry in the second set. Store the results